test_find_longest_permutation_random
